refactor: recipeCardコンポーネント周りのリファクタリング
